# Server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PWP - Movie Suggestions",
        "contact": {
            "name": " the backend developer",
            "email": "aujhazi20@student.oulu.fi"
        },
        "license": {
            "name": "Contact the frontend developer",
            "url": "mailto:hadi.mir@student.oulu.fi"
        },
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://18.196.196.249/",
            "description": "PWP Demo API Server"
        }
    ],
    "paths": {
        "/api/contents": {
            "get": {
                "tags": [
                    "Contents"
                ],
                "summary": "get the complete list of content available",
                "description": "Returns list of movies",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Contents"
                ],
                "summary": "add content",
                "description": "add content",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "trailer_url",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "director",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/contents/{content}": {
            "get": {
                "tags": [
                    "Content"
                ],
                "summary": "show content",
                "description": "show content",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "put": {
                "tags": [
                    "Content"
                ],
                "summary": "edit content",
                "description": "edit content",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "trailer_url",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "director",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "release_date",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Content"
                ],
                "summary": "delete content",
                "description": "delete content",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/contents/{content}/genres": {
            "get": {
                "tags": [
                    "Content Genres"
                ],
                "summary": "list of genres belonging to a certain content",
                "description": "Returns list of genres belonging to a certain content",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "put": {
                "tags": [
                    "Content Genres"
                ],
                "summary": "Update list of genres for selected content. Takes list of IDs.",
                "description": "Update list of genres for selected content. Takes list of IDs.",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "content",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/users/{user}/favorites": {
            "get": {
                "tags": [
                    "User Favorites"
                ],
                "summary": "list of favorite contents belonging to a certain user",
                "description": "Returns list of favorite contents belonging to a certain user",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "User Favorites"
                ],
                "summary": "Add Favorite",
                "description": "Adds the selected content to the list of the users' favorite contents.",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "favorite",
                        "in": "query",
                        "description": "content to be added as favorite",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/users/{user}/favorites/{favorite}": {
            "delete": {
                "tags": [
                    "User Favorite"
                ],
                "summary": "Remove favorite",
                "description": "From the list of a users favorite contents, remove the selected one.",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "favorite",
                        "in": "path",
                        "description": "content to be removed as favorite",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                }
            }
        },
        "/api/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "complete list of all genres",
                "description": "Returns list of genres",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "post": {
                "tags": [
                    "Genres"
                ],
                "summary": "add genre",
                "description": "add genre",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "409": {
                        "description": "Data exists"
                    }
                }
            }
        },
        "/api/genres/{genre}": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "show genre",
                "description": "show genre",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Genre"
                ],
                "summary": "delete genre",
                "description": "delete genre",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "genre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User actions"
                ],
                "summary": "user login",
                "description": "Gets the user ID and the user hash based on email and password.",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operation successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User actions"
                ],
                "summary": "user logout",
                "description": "remove the hash used for the users authentication",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/profiles": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create profile",
                "description": "Creates a new profile",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/profiles/{profile}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show profile",
                "description": "Get all info about a user profile",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update profile",
                "description": "Update the current users profile",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "birth_date",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "409": {
                        "description": "data exists"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "delete profile",
                "description": "soft delete the profile of a current user - the account will still continue to exist.",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "405": {
                        "description": "Method not allowed"
                    }
                }
            }
        },
        "/api/update-password": {
            "post": {
                "tags": [
                    "User actions"
                ],
                "summary": "change password",
                "description": "updates the password of the user logged in",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "new-password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "403": {
                        "description": "Access denied"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "409": {
                        "description": "data exists"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "User actions"
                ],
                "summary": "Register user",
                "description": "Register user",
                "operationId": "getProjectsList",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "409": {
                        "description": "data exists"
                    }
                }
            }
        }
    },
...
    "security": [
        []
    ]
}